Gestion_CETA/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.0
djangorestframework==3.15.1
django_filter==24.1
djangorestframework-jsonapi==6.1.0
djangorestframework-simplejwt==5.3.1
inflection==0.5.1
pip==24.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
reportlab==4.2.0